P31XX: libsensors: fix buffer overflows
long/int can't hold timestamp values in ns, use int64_t instead

+++ libsensors/gp2a_proximity.c
@@ -113,7 +113,7 @@ int gp2a_proximity_activate(struct piranha_sensors_handlers *handlers)
 
 	data = (struct gp2a_proximity_data *) handlers->data;
 
-	rc = sysfs_value_write(data->path_enable, 1);
+	rc = sysfs_value_write(data->path_enable, (int64_t) 1);
 	if (rc < 0) {
 		ALOGE("%s: Unable to write sysfs value", __func__);
 		return -1;
@@ -136,7 +136,7 @@ int gp2a_proximity_deactivate(struct piranha_sensors_handlers *handlers)
 
 	data = (struct gp2a_proximity_data *) handlers->data;
 
-	rc = sysfs_value_write(data->path_enable, 0);
+	rc = sysfs_value_write(data->path_enable, (int64_t) 0);
 	if (rc < 0) {
 		ALOGE("%s: Unable to write sysfs value", __func__);
 		return -1;
@@ -147,9 +147,9 @@ int gp2a_proximity_deactivate(struct piranha_sensors_handlers *handlers)
 	return 0;
 }
 
-int gp2a_proximity_set_delay(struct piranha_sensors_handlers *handlers, long int delay)
+int gp2a_proximity_set_delay(struct piranha_sensors_handlers *handlers, int64_t delay)
 {
-	ALOGD("%s(%p, %ld)", __func__, handlers, delay);
+	ALOGD("%s(%p, %" PRId64 ")", __func__, handlers, delay);
 
 	return 0;
 }
